pkp/pkp-lib#6145 move report generator out of tools
